[CIR] Move isTriviallyRecursive to MangleContext
Hosting isTriviallyRecursive on FunctionDecl pulled a MangleContext
dependency into an AST declaration, which is a layering violation.  Move
the predicate, along with its FunctionIsDirectlyRecursive body visitor,
onto MangleContext -- the mangled/asm-name comparison it performs is what
ties it there.  Classic CodeGen and CIRGen both call
getMangleContext().isTriviallyRecursive(FD); Decl.h no longer references
MangleContext.

Also collapse the test's separate OGCG check prefix into LLVM: the
CIR-lowered and classic -emit-llvm outputs are byte-identical for this
absence check, so one LLVM prefix runs on both files.
