+++ clang/include/clang/AST/Mangle.h
@@ -129,6 +129,14 @@ public:
   virtual bool shouldMangleCXXName(const NamedDecl *D) = 0;
   virtual bool shouldMangleStringLiteral(const StringLiteral *SL) = 0;
 
+  /// Return true if \p FD's body contains a direct call back to the symbol it
+  /// links as, through an asm label or a __builtin_* alias (PR9614 / glibc's
+  /// btowc pattern).  An available_externally body of such a function is not a
+  /// valid stand-in for the real implementation and should be dropped before
+  /// codegen.  The check keys off the mangled/asm name, hence it lives on
+  /// MangleContext.
+  bool isTriviallyRecursive(const FunctionDecl *FD);
+
   virtual bool isUniqueInternalLinkageDecl(const NamedDecl *ND) {
     return false;
   }
